+++ mpl_composite/__init__.py
@@ -4,6 +4,7 @@ from .axis import PlotAxis, ScaleLinear, ScaleLinLog, ScaleLog, Ticks
 from .elements import Composite, Spacer, Text
 from .figure import CompositeFigure, save_figure
 from .geometry import HAlign, Margin, Range, VAlign
+from .plot import PlotAxes
 from .style import ColorRange, FontWeight, LineStyle, TextStyle, Theme
 
 __all__ = [
@@ -14,6 +15,7 @@ __all__ = [
     "HAlign",
     "LineStyle",
     "Margin",
+    "PlotAxes",
     "PlotAxis",
     "Range",
     "ScaleLinLog",