+++ mpl_composite/plot/__init__.py
@@ -1,0 +1,5 @@
+"""A conventional plot, rebuilt as an element family."""
+
+from ._plot_axes import PlotAxes
+
+__all__ = ["PlotAxes"]