+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Classic x/y plot element family: framed plot area with grid, drawn axes with ticks and labels, title
+
 ### Changed
 
 ### Deprecated